requirements.txt
llama-index-core>=0.10.0
llama-index-readers-file>=0.1.4
llama-index-vector-stores-chroma>=0.1.1
llama-index-embeddings-google>=0.1.3
google-generativeai>=0.3.0
python-dotenv>=1.0.0
chromadb>=0.4.22
pypdf>=3.17.0